ICP code with feature in error function now compiles

+++ kidson/pclforSVN/transformation_estimation_feature_matches.h
@@ -81,7 +81,7 @@ namespace pcl
         	mPointsFeatures = numberPoints;
         }
 
-        virtual void setFeatureErrorWeight(int alpha)
+        virtual void setFeatureErrorWeight(float alpha)
         {
         	featureErrorWeight = alpha;
         }
